requirements.txt
langchain
langchain-community
openai
python-dotenv
streamlit
langchain-openai